# api/openapi/gateway.yaml
openapi: 3.0.3
info:
  title: Animus Gateway API
  version: 0.2.0
servers:
  - url: http://localhost:8080
paths:
...
  /api/lineage/{proxyPath}:
    parameters:
      - name: proxyPath
        in: path
        required: true
        schema:
          type: string
        description: Remaining path to proxy (may include slashes).
    get:
      summary: Proxy GET to Lineage
      tags: [Proxy]
      security:
        - bearerAuth: []
      responses:
        "200":
          $ref: "#/components/responses/ProxiedOK"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "502":
          $ref: "#/components/responses/BadGateway"
    post:
      summary: Proxy POST to Lineage
      tags: [Proxy]
      security:
        - bearerAuth: []
      responses:
        "200":
          $ref: "#/components/responses/ProxiedOK"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "502":
          $ref: "#/components/responses/BadGateway"
    put:
      summary: Proxy PUT to Lineage
      tags: [Proxy]
      security:
        - bearerAuth: []
      responses:
        "200":
          $ref: "#/components/responses/ProxiedOK"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "502":
          $ref: "#/components/responses/BadGateway"
    patch:
      summary: Proxy PATCH to Lineage
      tags: [Proxy]
      security:
        - bearerAuth: []
      responses:
        "200":
          $ref: "#/components/responses/ProxiedOK"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "502":
          $ref: "#/components/responses/BadGateway"
    delete:
      summary: Proxy DELETE to Lineage
      tags: [Proxy]
      security:
        - bearerAuth: []
      responses:
        "200":
          $ref: "#/components/responses/ProxiedOK"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "502":
          $ref: "#/components/responses/BadGateway"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    ProxiedOK:
      description: Proxied response from upstream service
      content:
        application/json:
          schema:
            type: object
            additionalProperties: true
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    BadGateway:
      description: Upstream error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    ErrorResponse:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          type: string
        request_id:
          type: string